Added some custom response examples

+++ db/db_user.py
@@ -22,14 +22,20 @@ def get_user(db: Session, id: int):
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        return False
     user.update({
         DbUser.username: request.username,
         DbUser.email: request.email,
         DbUser.password: Hash.bcrypt(request.password)
     })
     db.commit()
+    return True
 
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        return False
     user.delete()
     db.commit()
+    return True
